STDTC-71 suppress react-intl warnings in test harness
`react-intl` is really REALLY noisy on the console when it finds
uncompiled translations. This is useful in production where we're
worried about efficiency, but counter-productive in development where we
don't have access to the compiled translations (they're published with
releases but not committed to the repository). IOW, the warnings we get
about uncompiled translations during development/testing are totally
bogus, so this PR silences them.

Refs STDTC-71